reorganizing test folders
